code cleanup: organized swift files
