+++ WeatherPOV/Base.lproj/Main.storyboard
@@ -9,7 +9,7 @@
         <!--Data Source-->
         <scene sceneID="tne-QT-ifu">
             <objects>
-                <viewController id="BYZ-38-t0r" customClass="ViewController" customModule="WeatherPOV" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController storyboardIdentifier="WPForecastController" useStoryboardIdentifierAsRestorationIdentifier="YES" id="BYZ-38-t0r" customClass="WPForecastController" customModule="WeatherPOV" customModuleProvider="target" sceneMemberID="viewController">
                     <layoutGuides>
                         <viewControllerLayoutGuide type="top" id="y3c-jy-aDJ"/>
                         <viewControllerLayoutGuide type="bottom" id="wfy-db-euE"/>
